Don't overwrite existing value of $wgSyntaxHighlightModels
Currently extensions that want to add their model to $wgSyntaxHighlightModels
must do so in an extension function since they don't want to assume
that their extension is loaded after SyntaxHighlight GeSHi is.
But by not overwriting the array during initialization, extensions
don't have to worry about load order.

Also require a minimum of MW 1.24 which removed support for
register_globals.

Change-Id: I15c56690d156e90991aa86a3d79dbeaeb18dd6fd