Update patch set 2
Patch Set 2:

(1 comment)

> Are the commas at the end of [new] lines 54 & 138 needed?

No, they're optional. But it's nice to have them because they make the diff clearer if another element is added to the end of the array.

> Shouldn't the closing " ), " on [new] line 55 be over to the right a bit?

It is correctly indented by MediaWiki's standards. But comparing to other arrays in the file, GeSHi's standards appear to be different.

Patch-set: 2