Update patch set 5
Patch Set 5: Code-Review-1

I2ad43801cc7128 should not have been merged. This fixes the problem, but is still not an improvement compared to status quo in my opinion.

Exactly what is the problem with WikiApiary? There should be no reason to hardcode the version in the extension files. There was no problem to begin with. It's only declared in one file (geshi.php) and included at setup.

Patch-set: 5
Label: Code-Review=-1