Add api-pretty-content class when pretty-printing API output
The API has wrapped its pretty-printed output since Id9cdf102. Apply an
appropriate class to preserve this now that GeSHi is handling it.

Unfortunately GeSHi itself doesn't support adding more than one arbitrary class
to the <pre> (and we're already using that), so we have to add it in a
post-processing step.

Bug: T88742
Change-Id: I38e41db5c341fe06ff825c82d5a9cd4810b7cc24