Update patch set 2
Patch Set 2:

@Edsanders: do the client-side and server-side syntax highlighting implementations have to produce identical output?

if I we were to do-over syntax highlighting in MediaWiki, I would advocate for using a JavaScript library that could be run client-side as well as be set up as a backend nodejs service. But I have a limited amount of time to spend on this, so my goal is to leave things in a better state than I found them. Switching the backend to Pygments would solve T85794 and extend syntax highlighting support to a slew of newer languages that GeSHi does not support (like Go and Rust) while maintaining backward-compatibility.

Patch-set: 2