Update patch set 2
Patch Set 2:

In an very ideal world, yes. But if such a library (of sufficient quality) doesn't exist, then we can make compromises. Client side highlighting is also not a show-stopper as we can do Parsoid round-trips.

Currently with GeSHi we have to use resource loader to fetch the CSS if you change language mid-edit, so we definitely need to support fetching of language definitions asynchronously (assuming they aren't all loaded upfront).

Patch-set: 2