Update patch set 3
Patch Set 3:

In PS3 I added caching for the output of pygments, based on an md5 of the input. With that in place, I think the performance impact of this change will be positive on balance:

* 2 kB of JavaScript removed from the critical path of all page views.
  (Impact: every page view for all users).
* Improved cache hit rate for pages containing code blocks, due to there
  being only one module for highlighting all languages. (Impact: views of
  pages with highlighted code blocks for all users.)
* Equivalent performance when saving edits to pages that contain code
  blocks, when the edit does not touch the code.
* 100-200ms hit on edits that touch highlighted code blocks.
  (Impact: a fraction of edits on a fraction of articles.)

Apart from the performance benefits, there is of course the fact that a number of newer programming languages will now be supported.

Patch-set: 3