Update patch set 2
Patch Set 2:

The order of CSS is part of how CSS works. That's not a quirk, but an important feature to embrace. If it were any two files in general, I would agree the order is arbitrary and one should not rely on cascading order but be consistent (e.g. no need to overwrite your own values).

However, while in the same module, these two files are conceptually different. One is upstream, the other the local layer on top of that library. Similar to how user > skin > extension > skin > core styles. And other modules in general.

Patch-set: 2