Update patch set 1
Patch Set 1: Code-Review-2

Per T104109#1408867: The requirements for code that runs on the Wikimedia cluster are not merely that it behave well here and now, but that it continue to behave well even as it is ignored, misused, or substantially repurposed. The ratio of number of maintainers to code surface mandates being conservative in this way.

Patch-set: 1
Reviewer: Ori.livneh <255@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2