Update patch set 6
Patch Set 6:

Ok I thought there was. What was agreed on was to create a new extension call SyntaxHighlight and revert the old extension SyntaxHighlight_GeSHi to its old self. All this is to be done after reviewers have finished revieweing all the follow up patches to it. and This proposal will be looked at after 2 weeks.

Patch-set: 6