Revert "Don't top-load CSS" (causes FOUC)
This reverts commit 024d62be45f11f8ca3706887112726bcea914fb7.

Loading this in the bottom causes a flash of unstyled content
in Chrome on MediaWiki-namespace CSS/JS pages where the text is
first black un-bolded on a grey background and then re-arranges
with bolding in a few places (visually widening some pieces of
text and bumping them over). This is visually distracting and
somewhat confusing.

There are upcoming plans to collapse the stylesheet queue (and even
the dynamic load queue). Let's start with the ones that are actually
causing FOUCs like this one.

Impact should be minimal as it only affects pages where this
module is loaded.

Change-Id: Iaf1ce5e5b8cff56e987581a3dbe7f24460badde2