Update patch set 2
Abandoned

This won't work; the file we fetch from PyPI is not an ordinary zip file but a wheel file. The zipped up repo from bitbucket is not equivalent. To bundle the tip of master, one has to clone the repository and run `python setup.py bdist_wheel`.

Patch-set: 2
Status: abandoned