Preserve context when browser scrolls to L-num fragment
When following permalinks to individual lines of code,
e.g. https://en.wikipedia.org/wiki/MediaWiki:Common.js#L-9
... instead of the browser scrolling down until the selected line hits
the top edge at 0, ask the browser to stop a bit earlier, when the
selected line is at 20% from the top. This gives the reader some
context above/below the code in question.

While at it, undo a slight bit of extreme CSS obfuscation in the form
of the `&-lines` Less.js parent selector. There is little benefit to
de-compiling partial strings this way. Note that these are re-compiled
server-side before the browser sees it, so it does not make the
payload smaller. (And gzip does this in a way that naturally achieves
higher compression, already.)

Code is generally written for human readability, and applying this
gzip-esque compression by hand to the source code, does not appear to
benefit the maintainer. We don't do this in the PHP or JS code that
generates this HTML, either. So it is not like this represents or
derives from how the structure is actually developed. This helps
with greppability and Codesearch as well.

Change-Id: I5fc19d581e09e9b33b1019f7d21343487ab6f5c7